move known files rule to RuleTypes
